refactor: 파일 이름 수정 및 타입 수정
